requirements.txt
Pillow==8.3.0
qrcode==7.1
